Automatically convert most sources to python 3 AST-compatible
Run via:
    futurize -f print_with_import -f exec -f except -f raise \
             -f numliterals -f tuple_params -f paren -wn <files>

Where <files> are anything that this patch was run on before rebasing
later in the history. This means that this patch is no longer complete,
but also doesn't unfairly cover more than it used to for purposes of
branch comparison.

Any split __future__ imports have been combined.

This leaves a couple of problems in place that will be fixed separately:
- 8 xxx_todo_changeme placeholders from tuple parameter expansion
- Single instance of parentheses-less tuples
- Three-term raise syntax
- String raises
- Async keyword conflict